fix(dream): generate the eval-synth prompt grammar from the loader single source (#2646)
The LLM eval-synthesizer prompt hand-duplicated the loader's strict matcher
grammar. The earlier enrichment closed the original drop-every-candidate bug,
but the prompt and the loader stayed two independent copies of the operator set
and the matcher kinds — a copy that silently drifts the next time the loader
changes, re-opening the 100% candidate-drop.

Introduce a single source of truth (MATCHER_OPERATORS / MATCHER_KINDS in
teatree.eval.models): the loader compiles _OP_PATTERN from MATCHER_OPERATORS and
keys its expect-entry rejection on MATCHER_KINDS; the synthesizer prompt
enumerates both, so it can never instruct the model to emit a shape the loader
rejects. A parity test locks the coupling — a future operator/kind added to the
loader but missing from the prompt fails the test instead of dropping candidates.

Split the SDK-backed synthesizer (prompt + headless turn + reply parse) out of
llm_eval_proposer into sdk_eval_synthesizer so each module stays one concern and
under the module-health cap.
