+++ src/teatree/loops/dream/llm_eval_proposer.py
@@ -323,11 +323,15 @@ def stage_proposals_file(
         name = str(row.get("scenario_name") or "")
         if name:
             slices[name] = str(row.get("seed_citation") or "")
+    if synthesizer is None:
+        from teatree.loops.dream.sdk_eval_synthesizer import sdk_spec_synthesizer  # noqa: PLC0415
+
+        synthesizer = sdk_spec_synthesizer
     return stage_derived_evals(
         candidates,
         transcript_slices=slices,
         staging_dir=staging_dir or default_staging_dir(),
-        synthesizer=synthesizer or sdk_spec_synthesizer,
+        synthesizer=synthesizer,
         dry_run=dry_run,
     )
 
@@ -401,196 +405,12 @@ def _op(matcher: Matcher) -> str:
     return f'{matcher.operator} "{matcher.value}"'
 
 
-_SYNTH_SYSTEM_PROMPT = (
-    "You design ONE under_load behavioural eval that pins a drift rule. From the "
-    "drift rule, the cited real mistake, and a session slice, emit discriminating "
-    "matchers: a POSITIVE for the corrected behaviour and a NEGATIVE for the drift. "
-    "Use ONLY the existing matcher shapes; never invent a rule the slice cannot "
-    "ground. Also emit the cited drift's ACTUAL tool-call shape and the compliant "
-    "tool-call shape so the gate can prove your matchers reject the cited drift. "
-    "Reply with EXACTLY ONE JSON object and NO surrounding prose."
-)
-
-_SYNTH_PROMPT_TEMPLATE = (
-    "Design one under_load eval scenario as a SINGLE JSON object. "
-    "REQUIRED keys (emit EVERY one — a scenario missing any is dropped): "
-    "scenario_name (copy verbatim: {scenario_name}), context_preamble (a polluted "
-    "session prefix synthesized from the slice below), prompt (the user request that "
-    "triggers the drift), expect (a JSON list of matcher objects — see MATCHER "
-    'GRAMMAR below), fail_tool_call (a JSON object {{"name": <tool>, "input": '
-    "{{...}}}} for the cited DRIFT action your NEGATIVE matcher must reject), "
-    "pass_tool_call (the same shape for the COMPLIANT action your matchers must "
-    "accept). OPTIONAL keys (include when useful, safe to omit): scenario_description "
-    "(one sentence), agent_path (the owning skill, e.g. skills/rules/SKILL.md), "
-    "judge_rubric (a one-sentence PASS-iff rubric). "
-    "The matchers MUST reject fail_tool_call and accept pass_tool_call.\n\n"
-    "MATCHER GRAMMAR (the loader is STRICT — copy the SHAPE of each example exactly). "
-    'Every operator value is ALWAYS `contains "<substring>"` or `~ "<regex>"`:\n'
-    "  - positive tool_call — a `tool_call` key plus EXACTLY ONE `args.<path>` key "
-    "(no more, no fewer):\n"
-    '      {{"tool_call": "Bash", "args.command": "contains \\"git worktree add\\""}}\n'
-    "  - no_tool_call_matching — a single inner mapping holding EXACTLY ONE "
-    "`<tool>.<arg>` key (the key MUST contain a dot):\n"
-    '      {{"no_tool_call_matching": {{"Bash.command": "~ \\"rm -rf\\""}}}}\n'
-    "  - any_of — a NON-EMPTY list of positive `tool_call` entries ONLY (each itself "
-    "a single-`args.<path>` object):\n"
-    '      {{"any_of": [{{"tool_call": "Task", "args.prompt": "~ \\"fix\\""}}, '
-    '{{"tool_call": "Agent", "args.prompt": "~ \\"fix\\""}}]}}\n'
-    "  - final_state — one operator expression over the agent's FINAL message:\n"
-    '      {{"final_state": "~ \\"opened PR\\""}}\n'
-    "An expect entry that is none of these four kinds, a positive `tool_call` with "
-    "zero or several `args.<path>` keys, or a `no_tool_call_matching` with zero or "
-    "several inner entries is REJECTED and the whole scenario is dropped.\n\n"
-    "Reply with EXACTLY ONE JSON object and NO surrounding prose, markdown fences, "
-    "or trailing objects.\n\n"
-    "Drift rule: {drift_rule}\n"
-    "Cited real mistake: {seed_citation}\n\n"
-    "Session slice:\n{slice}"
-)
-
-_SYNTH_WATCHDOG_SECONDS = 5 * 60
-_SYNTH_MODEL = "claude-haiku-4-5"
-_REQUIRED_SYNTH_KEYS = ("scenario_name", "context_preamble", "prompt", "expect", "fail_tool_call", "pass_tool_call")
-
-
-def sdk_spec_synthesizer(candidate: Mapping[str, object], transcript_slice: str) -> SynthesizedSpec:
-    """The real synthesizer: one bounded headless SDK turn → a scenario, parsed defensively.
-
-    Mirrors :func:`teatree.loops.dream.sdk_distiller.sdk_distiller`'s invocation shape (the
-    ``claude_code`` preset, ``bypassPermissions``, a wall-clock watchdog) for a single
-    no-tool turn that transforms the candidate + slice into one scenario JSON object.
-    Raises on an unavailable ``claude`` or a malformed reply, so the caller DROPS the
-    candidate (never a staged unproven spec) rather than reporting a fake success.
-    """
-    import asyncio  # noqa: PLC0415
-    import shutil  # noqa: PLC0415
-
-    if shutil.which("claude") is None:
-        msg = "claude is not installed — the dream eval synthesizer cannot run"
-        raise RuntimeError(msg)
-    prompt = _SYNTH_PROMPT_TEMPLATE.format(
-        scenario_name=str(candidate.get("scenario_name") or ""),
-        drift_rule=str(candidate.get("drift_rule") or ""),
-        seed_citation=str(candidate.get("seed_citation") or ""),
-        slice=transcript_slice,
-    )
-    raw = asyncio.run(_collect_synth_turn(prompt))
-    return _parse_synthesized(raw, candidate)
-
-
-async def _collect_synth_turn(prompt: str) -> str:
-    import asyncio  # noqa: PLC0415
-
-    from claude_agent_sdk import AssistantMessage, ClaudeAgentOptions, ClaudeSDKClient, TextBlock  # noqa: PLC0415
-    from claude_agent_sdk.types import SystemPromptPreset  # noqa: PLC0415
-
-    options = ClaudeAgentOptions(
-        system_prompt=SystemPromptPreset(type="preset", preset="claude_code", append=_SYNTH_SYSTEM_PROMPT),
-        model=_SYNTH_MODEL,
-        permission_mode="bypassPermissions",
-        max_turns=1,
-        allowed_tools=[],
-    )
-    parts: list[str] = []
-    async with ClaudeSDKClient(options=options) as client:
-        await client.query(prompt)
-
-        async def _drain() -> list[object]:
-            return [message async for message in client.receive_response()]
-
-        for message in await asyncio.wait_for(_drain(), timeout=_SYNTH_WATCHDOG_SECONDS):
-            if isinstance(message, AssistantMessage):
-                parts.extend(block.text for block in message.content if isinstance(block, TextBlock))
-    return "\n".join(parts)
-
-
-def _extract_json_object(raw: str) -> Mapping[str, object] | None:
-    """The FIRST balanced JSON object in *raw*, tolerating prose and trailing objects.
-
-    The object analogue of :func:`teatree.loops.dream.sdk_distiller._extract_json_array`:
-    rather than spanning the first ``{`` to the last ``}`` (which captures multiple
-    objects or a trailing fragment and makes ``json.loads`` raise ``Extra data``), it
-    scans each ``{`` with :meth:`json.JSONDecoder.raw_decode` and returns the first
-    that decodes to a mapping — so a reply carrying prose plus more than one object
-    yields its first object instead of crashing the whole derivation phase.
-    """
-    import json  # noqa: PLC0415
-
-    decoder = json.JSONDecoder()
-    index = raw.find("{")
-    while index != -1:
-        try:
-            parsed, _ = decoder.raw_decode(raw, index)
-        except json.JSONDecodeError:
-            index = raw.find("{", index + 1)
-        else:
-            return parsed
-    return None
-
-
-def _parse_synthesized(raw: str, candidate: Mapping[str, object]) -> SynthesizedSpec:
-    """Parse the synthesizer's JSON object into a :class:`SynthesizedSpec`.
-
-    Extracts the first balanced JSON object (via :func:`_extract_json_object`), so
-    surrounding prose or a trailing second object no longer raises ``Extra data``. A
-    missing required key or a non-list ``expect`` raises, so a malformed reply DROPS
-    the candidate rather than staging a partial scenario.
-    """
-    payload = _extract_json_object(raw)
-    if payload is None:
-        msg = "synthesizer returned no JSON object"
-        raise ValueError(msg)
-    missing = [key for key in _REQUIRED_SYNTH_KEYS if key not in payload]
-    if missing:
-        msg = f"synthesized scenario is missing required key(s): {', '.join(missing)}"
-        raise ValueError(msg)
-    raw_expect = payload["expect"]
-    if not isinstance(raw_expect, list) or not raw_expect:
-        msg = "synthesized scenario has no matchers"
-        raise ValueError(msg)
-    matchers: list[Mapping[str, object]] = [
-        {str(key): value for key, value in entry.items()} for entry in raw_expect if isinstance(entry, Mapping)
-    ]
-    fail_tool_call = _require_tool_call(payload["fail_tool_call"], "fail_tool_call")
-    pass_tool_call = _require_tool_call(payload["pass_tool_call"], "pass_tool_call")
-    return SynthesizedSpec(
-        scenario_name=str(payload["scenario_name"] or candidate.get("scenario_name") or ""),
-        scenario_description=str(payload.get("scenario_description") or ""),
-        agent_path=str(payload.get("agent_path") or "skills/rules/SKILL.md"),
-        context_preamble=str(payload["context_preamble"]),
-        prompt=str(payload["prompt"]),
-        expect=matchers,
-        fail_tool_call=fail_tool_call,
-        pass_tool_call=pass_tool_call,
-        judge_rubric=str(payload.get("judge_rubric") or ""),
-    )
-
-
-def _require_tool_call(value: object, key: str) -> ToolCallShape:
-    """Validate a synthesized tool-call shape (``{"name": <tool>, "input": {...}}``).
-
-    The teeth check seeds its candidate-derived ``_fail`` / ``_pass`` transcripts
-    from these, so a malformed shape (no ``name``) is fatal — the candidate DROPS
-    rather than teeth-checking against an empty transcript that fails nothing.
-    """
-    if not isinstance(value, Mapping):
-        value = {}
-    fields = {str(field_key): field_value for field_key, field_value in value.items()}
-    name = str(fields.get("name") or "")
-    if not name:
-        msg = f"synthesized scenario has a malformed {key} (need a tool-call with a name)"
-        raise ValueError(msg)
-    tool_input = fields.get("input")
-    return ToolCallShape(name=name, input=dict(tool_input) if isinstance(tool_input, Mapping) else {})
-
-
 __all__ = [
     "DerivationOutcome",
     "SpecSynthesizer",
     "SynthesizedSpec",
     "default_staging_dir",
     "derive_eval_from_candidate",
-    "sdk_spec_synthesizer",
     "stage_derived_evals",
     "stage_proposals_file",
 ]
